Move env tests to new file
